+++ openspec/changes/complete-reverse-sync/tasks.md
@@ -131,9 +131,10 @@
   - [x] candidate 준비, roundtrip 검증, local proof/manifest 생성을
     `reverse_sync/verification_service.py`로 분리하고 CLI는 runtime dependency를
     주입하는 호환 adapter만 유지합니다.
-  - [ ] batch 대상 선택, online snapshot 준비, publisher 호출의 state 전환을
-    package service로 분리하고 CLI에는 argument parsing, display, orchestration만
-    유지합니다.
+  - [x] explicit manifest integrity/identity 검증, publisher 호출, semantic
+    postcondition, backup 생성을 `reverse_sync/publish_service.py`로 분리합니다.
+  - [ ] batch 대상 선택과 online snapshot 준비의 state 전환을 package service로
+    분리하고 CLI에는 argument parsing, display, orchestration만 유지합니다.
 - [x] online verify(`push --dry-run`) 출력에 run ID, base version/hash, local gates,
   push eligibility, reason code를 표시합니다.
 - [x] `push`가 explicit run/manifest를 받도록 합니다.
